+++ compiler-utils/src/main/java/com/squareup/anvil/compiler/internal/reference/MemberFunctionReference.kt
@@ -1,34 +1,67 @@
 package com.squareup.anvil.compiler.internal.reference
 
 import com.squareup.anvil.annotations.ExperimentalAnvilApi
-import com.squareup.anvil.compiler.internal.reference.TopLevelFunctionReference.Descriptor
-import com.squareup.anvil.compiler.internal.reference.TopLevelFunctionReference.Psi
+import com.squareup.anvil.compiler.internal.reference.MemberFunctionReference.Descriptor
+import com.squareup.anvil.compiler.internal.reference.MemberFunctionReference.Psi
 import com.squareup.anvil.compiler.internal.reference.Visibility.INTERNAL
 import com.squareup.anvil.compiler.internal.reference.Visibility.PRIVATE
 import com.squareup.anvil.compiler.internal.reference.Visibility.PROTECTED
 import com.squareup.anvil.compiler.internal.reference.Visibility.PUBLIC
 import com.squareup.anvil.compiler.internal.requireFqName
+import org.jetbrains.kotlin.descriptors.ClassConstructorDescriptor
 import org.jetbrains.kotlin.descriptors.DescriptorVisibilities
 import org.jetbrains.kotlin.descriptors.FunctionDescriptor
+import org.jetbrains.kotlin.descriptors.Modality.ABSTRACT
 import org.jetbrains.kotlin.lexer.KtTokens
+import org.jetbrains.kotlin.lexer.KtTokens.ABSTRACT_KEYWORD
+import org.jetbrains.kotlin.lexer.KtTokens.PUBLIC_KEYWORD
 import org.jetbrains.kotlin.name.FqName
+import org.jetbrains.kotlin.name.Name
+import org.jetbrains.kotlin.psi.KtConstructor
 import org.jetbrains.kotlin.psi.KtFunction
+import org.jetbrains.kotlin.psi.KtNamedFunction
 import org.jetbrains.kotlin.psi.psiUtil.visibilityModifierTypeOrDefault
 import org.jetbrains.kotlin.resolve.descriptorUtil.fqNameSafe
 import kotlin.LazyThreadSafetyMode.NONE
 
+/**
+ * Used to create a common type between [KtNamedFunction] class references and
+ * [FunctionDescriptor] references, to streamline parsing.
+ */
 @ExperimentalAnvilApi
-public sealed class TopLevelFunctionReference : AnnotatedReference, FunctionReference {
+public sealed class MemberFunctionReference : AnnotatedReference, FunctionReference {
+
+  public abstract val declaringClass: ClassReference
+
+  public override val module: AnvilModuleDescriptor get() = declaringClass.module
 
   protected abstract val returnType: TypeReference?
 
   public override fun returnTypeOrNull(): TypeReference? = returnType
 
+  public abstract fun isAbstract(): Boolean
+  public abstract fun isConstructor(): Boolean
+  public fun resolveGenericReturnTypeOrNull(
+    implementingClass: ClassReference
+  ): ClassReference? {
+    return returnType
+      ?.resolveGenericTypeOrNull(implementingClass)
+      ?.asClassReferenceOrNull()
+  }
+
+  public fun resolveGenericReturnType(implementingClass: ClassReference): ClassReference =
+    resolveGenericReturnTypeOrNull(implementingClass)
+      ?: throw AnvilCompilationExceptionFunctionReference(
+        functionReference = this,
+        message = "Unable to resolve return type for function $fqName with the implementing " +
+          "class ${implementingClass.fqName}."
+      )
+
   override fun toString(): String = "$fqName()"
 
   override fun equals(other: Any?): Boolean {
     if (this === other) return true
-    if (other !is TopLevelFunctionReference) return false
+    if (other !is MemberFunctionReference) return false
 
     if (fqName != other.fqName) return false
 
@@ -41,9 +74,9 @@ public sealed class TopLevelFunctionReference : AnnotatedReference, FunctionRefe
 
   public class Psi internal constructor(
     public override val function: KtFunction,
-    override val fqName: FqName,
-    override val module: AnvilModuleDescriptor,
-  ) : TopLevelFunctionReference(), FunctionReference.Psi {
+    override val declaringClass: ClassReference.Psi,
+    override val fqName: FqName
+  ) : MemberFunctionReference(), FunctionReference.Psi {
 
     override val annotations: List<AnnotationReference.Psi> by lazy(NONE) {
       function.annotationEntries.map {
@@ -52,21 +85,27 @@ public sealed class TopLevelFunctionReference : AnnotatedReference, FunctionRefe
     }
 
     override val returnType: TypeReference.Psi? by lazy(NONE) {
-      function.typeReference?.toTypeReference(declaringClass = null, module)
+      function.typeReference?.toTypeReference(declaringClass, module)
     }
 
     override val parameters: List<ParameterReference.Psi> by lazy(NONE) {
       function.valueParameters.map { it.toParameterReference(this) }
     }
 
+    override fun isAbstract(): Boolean =
+      function.hasModifier(ABSTRACT_KEYWORD) ||
+        (declaringClass.isInterface() && !function.hasBody())
+
+    override fun isConstructor(): Boolean = function is KtConstructor<*>
+
     override fun visibility(): Visibility {
       return when (val visibility = function.visibilityModifierTypeOrDefault()) {
-        KtTokens.PUBLIC_KEYWORD -> PUBLIC
+        PUBLIC_KEYWORD -> PUBLIC
         KtTokens.INTERNAL_KEYWORD -> INTERNAL
         KtTokens.PROTECTED_KEYWORD -> PROTECTED
         KtTokens.PRIVATE_KEYWORD -> PRIVATE
-        else -> throw AnvilCompilationExceptionFunctionReference(
-          functionReference = this,
+        else -> throw AnvilCompilationExceptionClassReference(
+          classReference = declaringClass,
           message = "Couldn't get visibility $visibility for function $fqName."
         )
       }
@@ -75,9 +114,9 @@ public sealed class TopLevelFunctionReference : AnnotatedReference, FunctionRefe
 
   public class Descriptor internal constructor(
     public override val function: FunctionDescriptor,
-    override val fqName: FqName = function.fqNameSafe,
-    override val module: AnvilModuleDescriptor,
-  ) : TopLevelFunctionReference(), FunctionReference.Descriptor {
+    override val declaringClass: ClassReference.Descriptor,
+    override val fqName: FqName = function.fqNameSafe
+  ) : MemberFunctionReference(), FunctionReference.Descriptor {
 
     override val annotations: List<AnnotationReference.Descriptor> by lazy(NONE) {
       function.annotations.map {
@@ -90,8 +129,12 @@ public sealed class TopLevelFunctionReference : AnnotatedReference, FunctionRefe
     }
 
     override val returnType: TypeReference.Descriptor? by lazy(NONE) {
-      function.returnType?.toTypeReference(declaringClass = null, module)
+      function.returnType?.toTypeReference(declaringClass, module)
     }
+
+    override fun isAbstract(): Boolean = function.modality == ABSTRACT
+
+    override fun isConstructor(): Boolean = function is ClassConstructorDescriptor
 
     override fun visibility(): Visibility {
       return when (val visibility = function.visibility) {
@@ -99,8 +142,8 @@ public sealed class TopLevelFunctionReference : AnnotatedReference, FunctionRefe
         DescriptorVisibilities.INTERNAL -> INTERNAL
         DescriptorVisibilities.PROTECTED -> PROTECTED
         DescriptorVisibilities.PRIVATE -> PRIVATE
-        else -> throw AnvilCompilationExceptionFunctionReference(
-          functionReference = this,
+        else -> throw AnvilCompilationExceptionClassReference(
+          classReference = declaringClass,
           message = "Couldn't get visibility $visibility for function $fqName."
         )
       }
@@ -109,15 +152,21 @@ public sealed class TopLevelFunctionReference : AnnotatedReference, FunctionRefe
 }
 
 @ExperimentalAnvilApi
-public fun KtFunction.toTopLevelFunctionReference(
-  module: AnvilModuleDescriptor,
+public fun KtFunction.toFunctionReference(
+  declaringClass: ClassReference.Psi
 ): Psi {
-  return Psi(function = this, fqName = requireFqName(), module = module)
+  val fqName = if (this is KtConstructor<*>) {
+    declaringClass.fqName.child(Name.identifier("<init>"))
+  } else {
+    requireFqName()
+  }
+
+  return Psi(this, declaringClass, fqName)
 }
 
 @ExperimentalAnvilApi
-public fun FunctionDescriptor.toTopLevelFunctionReference(
-  module: AnvilModuleDescriptor,
+public fun FunctionDescriptor.toFunctionReference(
+  declaringClass: ClassReference.Descriptor,
 ): Descriptor {
-  return Descriptor(function = this, module = module)
+  return Descriptor(this, declaringClass)
 }
