+++ compiler/src/main/java/com/squareup/anvil/compiler/codegen/dagger/ProvidesMethodFactoryGenerator.kt
@@ -12,10 +12,10 @@ import com.squareup.anvil.compiler.daggerProvidesFqName
 import com.squareup.anvil.compiler.internal.buildFile
 import com.squareup.anvil.compiler.internal.capitalize
 import com.squareup.anvil.compiler.internal.reference.AnnotatedReference
-import com.squareup.anvil.compiler.internal.reference.AnvilCompilationExceptionFunctionalReference
+import com.squareup.anvil.compiler.internal.reference.AnvilCompilationExceptionFunctionReference
 import com.squareup.anvil.compiler.internal.reference.ClassReference
-import com.squareup.anvil.compiler.internal.reference.FunctionReference
-import com.squareup.anvil.compiler.internal.reference.PropertyReference
+import com.squareup.anvil.compiler.internal.reference.MemberFunctionReference
+import com.squareup.anvil.compiler.internal.reference.MemberPropertyReference
 import com.squareup.anvil.compiler.internal.reference.TypeReference
 import com.squareup.anvil.compiler.internal.reference.Visibility.INTERNAL
 import com.squareup.anvil.compiler.internal.reference.asClassName
@@ -299,11 +299,11 @@ internal class ProvidesMethodFactoryGenerator : PrivateCodeGenerator() {
 
   private fun checkFunctionIsNotAbstract(
     clazz: ClassReference.Psi,
-    function: FunctionReference.Psi
+    function: MemberFunctionReference.Psi
   ) {
-    fun fail(): Nothing = throw AnvilCompilationExceptionFunctionalReference(
+    fun fail(): Nothing = throw AnvilCompilationExceptionFunctionReference(
       message = "@Provides methods cannot be abstract",
-      functionalReference = function
+      functionReference = function
     )
 
     // If the function is abstract, then it's an error.
@@ -331,8 +331,8 @@ internal class ProvidesMethodFactoryGenerator : PrivateCodeGenerator() {
   }
 
   private class CallableReference(
-    private val function: FunctionReference.Psi? = null,
-    private val property: PropertyReference.Psi? = null
+    private val function: MemberFunctionReference.Psi? = null,
+    private val property: MemberPropertyReference.Psi? = null
   ) {
 
     init {
@@ -358,10 +358,10 @@ internal class ProvidesMethodFactoryGenerator : PrivateCodeGenerator() {
       function?.parameters?.mapToConstructorParameters() ?: emptyList()
 
     val type: TypeReference = function?.let {
-      it.returnTypeOrNull() ?: throw AnvilCompilationExceptionFunctionalReference(
+      it.returnTypeOrNull() ?: throw AnvilCompilationExceptionFunctionReference(
         message = "Dagger provider methods must specify the return type explicitly when using " +
           "Anvil. The return type cannot be inferred implicitly.",
-        functionalReference = it
+        functionReference = it
       )
     } ?: property!!.type()
     val annotationReference: AnnotatedReference = function ?: property!!
